feat: prepare map layout for next install
