+++ dash-ui/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "dayjs": "^1.11.10",
+    "maplibre-gl": "^3.6.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.22.3"
