+++ dash-ui/src/styles/global.css
@@ -51,43 +51,31 @@ main {
   flex: 1;
 }
 
-.app-shell {
-  position: relative;
+.layout-root {
   display: flex;
   width: 100vw;
   height: 100vh;
   overflow: hidden;
+  background: #000;
+  color: #f8fafc;
 }
 
-
-.app-shell__map {
+.layout-map {
   flex: 1;
   min-width: 0;
   min-height: 0;
   height: 100%;
-  width: 100%;
   display: flex;
-  background: var(--panel-bg);
-  border-right: 1px solid rgba(255, 255, 255, 0.04);
   position: relative;
-  z-index: 0;
 }
 
-.app-shell__aside {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  width: clamp(520px, 34vw, 680px);
-  max-width: 100%;
+.layout-aside {
+  height: 100%;
+  width: clamp(400px, 24vw, 520px);
   border-left: 1px solid rgba(255, 255, 255, 0.1);
   background: rgba(0, 0, 0, 0.35);
   backdrop-filter: blur(14px);
-  pointer-events: none;
   display: flex;
-  padding: clamp(16px, 2.4vw, 32px);
-  box-sizing: border-box;
-  z-index: 1;
 }
 
 .geo-scope-map {
@@ -101,6 +89,20 @@ main {
 .geo-scope-map__canvas {
   width: 100%;
   height: 100%;
+}
+
+.geo-scope-map__attribution {
+  position: absolute;
+  right: 16px;
+  bottom: 14px;
+  font-size: 12px;
+  line-height: 1.4;
+  color: rgba(255, 255, 255, 0.6);
+  background: rgba(0, 0, 0, 0.45);
+  padding: 4px 10px;
+  border-radius: 9999px;
+  letter-spacing: 0.02em;
+  pointer-events: none;
 }
 
 .geo-scope-map__fallback {
@@ -133,7 +135,6 @@ main {
   display: flex;
   flex-direction: column;
   gap: clamp(20px, 2.6vw, 36px);
-  pointer-events: none;
   overflow: hidden;
 }
