+++ dash-ui/tsconfig.json
@@ -12,9 +12,6 @@
     "module": "ESNext",
     "moduleResolution": "Node",
     "baseUrl": "./",
-    "paths": {
-      "maplibre-gl": ["src/vendor/maplibre-gl"]
-    },
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
